+++ packages/persistance-redis/src/repository/redis.repository.service.ts
@@ -6,7 +6,9 @@ import {
   ApplicationError,
   ConfigProviderService,
   Constants as CoreConstants,
-  getNested
+  GenericObject,
+  getNested,
+  setNested
 } from '@node-c/core';
 
 import { ValidationSchema, registerSchema, validate } from 'class-validator';
@@ -16,6 +18,8 @@ import { v4 as uuid } from 'uuid';
 import {
   EntitySchema,
   EntitySchemaColumnType,
+  FilterItemOptions,
+  GetValuesFromResultsOptions,
   PrepareOptions,
   RepositoryFindOptions,
   RepositoryFindPrivateOptions,
@@ -33,9 +37,12 @@ import { RedisStoreService } from '../store';
 export class RedisRepositoryService<Entity> {
   protected _columnNames: string[];
   protected _innerPrimaryKeys: string[];
+  protected _innerPrimaryKeysMap: GenericObject<boolean>;
   protected _primaryKeys: string[];
+  protected _primaryKeysMap: GenericObject<boolean>;
   protected defaultTTL?: number;
   protected defaultIndividualSearchEnabled: boolean;
+  protected experimentalDeletionEnabled: boolean = false;
   protected storeDelimiter: string;
   protected validationSchemaProperties: ValidationSchema['properties'];
   protected validationSettings: AppConfigCommonPersistanceNoSQLValidationSettings;
@@ -46,11 +53,17 @@ export class RedisRepositoryService<Entity> {
   public get innerPrimaryKeys(): string[] {
     return this._innerPrimaryKeys;
   }
+  public get innerPrimaryKeysMap(): GenericObject<boolean> {
+    return this._innerPrimaryKeysMap;
+  }
   public get persistanceModuleName(): string {
     return this._persistanceModuleName;
   }
   public get primaryKeys(): string[] {
     return this._primaryKeys;
+  }
+  public get primaryKeysMap(): GenericObject<boolean> {
+    return this._primaryKeysMap;
   }
 
   constructor(
@@ -67,7 +80,9 @@ export class RedisRepositoryService<Entity> {
     const { columns, name: entityName } = schema;
     const columnNames: string[] = [];
     const innerPrimaryKeys: string[] = [];
+    const innerPrimaryKeysMap: GenericObject<boolean> = {};
     const primaryKeys: string[] = [];
+    const primaryKeysMap: GenericObject<boolean> = {};
     const validationSchemaProperties: ValidationSchema['properties'] = {};
     for (const columnName in columns) {
       const { isInnerPrimary, primary, primaryOrder, validationProperties } = columns[columnName];
@@ -79,9 +94,10 @@ export class RedisRepositoryService<Entity> {
           );
         }
         primaryKeys.push(columnName);
-      }
-      if (isInnerPrimary) {
+        primaryKeysMap[columnName] = true;
+      } else if (isInnerPrimary) {
         innerPrimaryKeys.push(columnName);
+        innerPrimaryKeysMap[columnName] = true;
       }
       if (validationProperties) {
         validationSchemaProperties[columnName] = validationProperties;
@@ -89,9 +105,11 @@ export class RedisRepositoryService<Entity> {
     }
     this._columnNames = columnNames;
     this._innerPrimaryKeys = innerPrimaryKeys;
+    this._innerPrimaryKeysMap = innerPrimaryKeysMap;
     this._primaryKeys = primaryKeys.sort(
       (columnName0, columnName1) => columns[columnName0].primaryOrder! - columns[columnName1].primaryOrder!
     );
+    this._primaryKeysMap = primaryKeysMap;
     this.defaultTTL = settingsPerEntity?.[entityName]?.ttl || defaultTTL;
     if (typeof settingsPerEntity?.[entityName]?.defaultIndividualSearchEnabled !== 'undefined') {
       this.defaultIndividualSearchEnabled = settingsPerEntity?.[entityName]?.defaultIndividualSearchEnabled;
@@ -105,25 +123,25 @@ export class RedisRepositoryService<Entity> {
     registerSchema({ name: entityName, properties: validationSchemaProperties });
   }
 
-  // TODO: reduce the large numbers of whole-array iterations on the found results
+  // protected async delete
+
   async find<ResultItem extends Entity | string = Entity>(
     options: RepositoryFindOptions,
     privateOptions?: RepositoryFindPrivateOptions
   ): Promise<{ items: ResultItem[]; more: boolean }> {
     const { primaryKeys, schema, store, storeDelimiter } = this;
-    const { isArray, name: entityName, nestedObjectContainerPath, storeKey: entityStoreKey } = schema;
-    const { filters, findAll, page, perPage, individualSearch, withValues: optWithValues } = options;
+    const { name: entityName, storeKey: entityStoreKey } = schema;
+    const { filters, findAll, individualSearch, withValues: optWithValues } = options;
     const { requirePrimaryKeys } = privateOptions || {};
     const individualSearchEnabled =
       typeof individualSearch !== 'undefined' ? individualSearch : this.defaultIndividualSearchEnabled;
-    const paginationOptions: { count?: number; cursor?: number } = {};
-    const primaryKeyFiltersToForceCheck: string[] = [];
+    const primaryKeyFiltersToForceCheck: GenericObject<boolean> = {};
     const storeEntityKeys: string[] = [];
     const withValues = typeof optWithValues === 'undefined' || optWithValues === true ? true : false;
-    let count: number = 0;
     let hasNonPrimaryKeyFilters = false;
     let primaryKeyFiltersCount = 0;
     if (filters && Object.keys(filters).length) {
+      // set up the construction of the store keys by primary keys
       storeEntityKeys.push('');
       primaryKeys.forEach(field => {
         const value = filters[field];
@@ -163,7 +181,7 @@ export class RedisRepositoryService<Entity> {
               }
             } else {
               hasNonPrimaryKeyFilters = true;
-              primaryKeyFiltersToForceCheck.push(field);
+              primaryKeyFiltersToForceCheck[field] = true;
               storeEntityKeys[0] += `${storeDelimiter}*`;
               return;
             }
@@ -187,19 +205,17 @@ export class RedisRepositoryService<Entity> {
       if (!hasNonPrimaryKeyFilters) {
         hasNonPrimaryKeyFilters = primaryKeyFiltersCount === Object.keys(filters).length;
       }
-      if (!findAll) {
-        count = perPage || 100;
-        paginationOptions.count = count;
-        paginationOptions.cursor = (page ? page - 1 : 0) * count;
-      }
     } else if (!findAll) {
       throw new ApplicationError(
         `[RedisRepositoryService ${entityName}][Error]: ` +
           'Either filters or findAll is required when calling the find method.'
       );
     }
+    // findAll logic:
+    // if doing an inidividual search, go through the store keys one by one; all PKs are required;
+    // if doing a wildcard search, iterate a cursor through the whole database until returned to the start
     if (findAll) {
-      if (individualSearchEnabled && !primaryKeyFiltersCount) {
+      if (individualSearchEnabled && !primaryKeyFiltersCount && primaryKeys.length) {
         throw new ApplicationError(
           `[RedisRepositoryService ${entityName}][Error]: ` +
             'Primary key filters are required when findAll and individualSearchEnabled ' +
@@ -209,78 +225,62 @@ export class RedisRepositoryService<Entity> {
       let initialResults: ResultItem[] = [];
       // get the base results
       if (individualSearchEnabled) {
-        initialResults = (
-          await Promise.all(storeEntityKeys.map(key => store.get(`${entityStoreKey}${key}`, { parseToJSON: true })))
-        ).filter(item => item !== null) as ResultItem[];
+        initialResults = (await Promise.all(
+          storeEntityKeys.map(key => store.get(`${entityStoreKey}${key}`, { parseToJSON: true }))
+        )) as ResultItem[];
       } else {
         // TODO: if no filters are provided, this will not return anything
+        // TODO: (reply, some point later) WDYIM, Rumen?
         const scanData = await store.scan(`${entityStoreKey}${storeEntityKeys[0]}`, {
-          ...paginationOptions,
           parseToJSON: true,
           scanAll: findAll,
           withValues
         });
         initialResults = scanData.values as ResultItem[];
       }
-      if (nestedObjectContainerPath) {
-        initialResults = initialResults
-          .map(item => {
-            if (item && typeof item === 'object' && !(item instanceof Date)) {
-              return getNested(item, nestedObjectContainerPath).unifiedValue;
-            }
-            return item;
-          })
-          .filter(item => typeof item !== 'undefined' && item !== null) as ResultItem[];
-      }
-      if (isArray) {
-        initialResults = initialResults.flat() as ResultItem[];
-      }
-      if (!hasNonPrimaryKeyFilters) {
-        return { items: initialResults, more: false };
-      }
-      // filter by the results' object data
+      // filter the base results by inner keys, as well as retrieve items from arrays and nested items
       return {
-        items: initialResults.filter(item => {
-          let filterResult = true;
-          for (const key in filters) {
-            if (primaryKeys.includes(key) && !primaryKeyFiltersToForceCheck.includes(key)) {
-              continue;
-            }
-            const filterValue = filters[key];
-            const itemValue = (item as Record<string, unknown>)[key];
-            if (filterValue instanceof Array) {
-              if (!filterValue.includes(itemValue)) {
-                filterResult = false;
-                break;
-              }
-              continue;
-            }
-            if (filterValue !== itemValue) {
-              filterResult = false;
-              break;
-            }
-          }
-          return filterResult;
-        }) as ResultItem[],
+        items: this.getValuesFromResults(initialResults, {
+          filters,
+          hasNonPrimaryKeyFilters,
+          primaryKeyFiltersToForceCheck
+        }),
         more: false
       };
     }
-    // process non-findAll
+    // non-findAll logic:
+    // if doing an inidividual search, go through the store keys one by one; all PKs are required;
+    // if doing a wildcard search, iterate a cursor through the whole database until the pagination end is reached
+    const { page, perPage } = options || {};
+    // for non-individual search, we'll only have the first key anyway
+    // TODO: check whether the above is true nad apply it to findAll if it isn't
     const [storeEntityKey] = storeEntityKeys;
+    const count: number = perPage || 100;
+    const limit = count + 1;
+    let cursor = (page ? page - 1 : 0) * count;
     let more = false;
     let results: ResultItem[] = [];
-    while (results.length < count) {
+    while (results.length < limit) {
       let endReached = false;
       let iterationResults: ResultItem[] = [];
       // get the base results
       if (individualSearchEnabled) {
-        iterationResults = (
-          await Promise.all(storeEntityKeys.map(key => store.get(`${entityStoreKey}${key}`, { parseToJSON: true })))
-        ).filter(item => item !== null) as ResultItem[];
-        endReached = true;
+        const iterationLimit = cursor + limit;
+        const iterationPromises: Promise<ResultItem>[] = [];
+        for (let i = cursor; i < iterationLimit; i++) {
+          const key = storeEntityKeys[i];
+          if (!key) {
+            endReached = true;
+            break;
+          }
+          iterationPromises.push(store.get(`${entityStoreKey}${key}`, { parseToJSON: true }));
+        }
+        iterationResults = (await Promise.all(iterationPromises)) as ResultItem[];
+        cursor = iterationLimit;
       } else {
         const { cursor: newCursor, values: innerResults } = await store.scan(`${entityStoreKey}${storeEntityKey}`, {
-          ...paginationOptions,
+          count,
+          cursor,
           parseToJSON: true,
           scanAll: false,
           withValues
@@ -289,42 +289,24 @@ export class RedisRepositoryService<Entity> {
         if (newCursor === 0) {
           endReached = true;
         } else {
-          paginationOptions.cursor = newCursor;
+          cursor = newCursor;
         }
       }
-      if (nestedObjectContainerPath) {
-        iterationResults = iterationResults
-          .map(item => {
-            if (item && typeof item === 'object' && !(item instanceof Date)) {
-              return getNested(item, nestedObjectContainerPath).unifiedValue;
-            }
-            return item;
-          })
-          .filter(item => typeof item !== 'undefined' && item !== null) as ResultItem[];
-      }
-      if (isArray) {
-        iterationResults = iterationResults.flat() as ResultItem[];
-      }
-      // filter by the results' object data
+      // filter the base results by inner keys, as well as retrieve items from arrays and nested items;
+      // the beauty of this approach is that it follows the rules of the pagination regardless of whether
+      // the results are coming from nested items or not
       results = results.concat(
-        iterationResults.filter(item => {
-          let filterResult = true;
-          for (const key in filters) {
-            if (primaryKeys.includes(key)) {
-              continue;
-            }
-            if ((item as Record<string, unknown>)[key] !== filters[key]) {
-              filterResult = false;
-              break;
-            }
-          }
-          return filterResult;
-        }) as ResultItem[]
+        this.getValuesFromResults(iterationResults, {
+          filters,
+          hasNonPrimaryKeyFilters,
+          primaryKeyFiltersToForceCheck
+        })
       );
       if (endReached) {
         break;
       }
     }
+    // determine whether this is the end of the pagination
     if (results.length > count) {
       more = true;
       results = results.slice(0, count);
@@ -332,7 +314,67 @@ export class RedisRepositoryService<Entity> {
     return { items: results, more };
   }
 
-  // TODO: isArray support
+  protected filterItem<Item>(item: Item, filters: GenericObject<unknown>, options?: FilterItemOptions): boolean {
+    if (typeof item === 'undefined' || item === null) {
+      return false;
+    }
+    const { keysToSkip, skippableKeysToForceCheck } = options || {};
+    let filterResult = true;
+    for (const key in filters) {
+      if (keysToSkip?.[key] && !skippableKeysToForceCheck?.[key]) {
+        continue;
+      }
+      const filterValue = filters[key];
+      const itemValue = (item as GenericObject<unknown>)[key];
+      if (filterValue instanceof Array) {
+        if (!filterValue.includes(itemValue)) {
+          filterResult = false;
+          break;
+        }
+        continue;
+      }
+      // TODO: filter operators
+      if (filterValue !== itemValue) {
+        filterResult = false;
+        break;
+      }
+    }
+    return filterResult;
+  }
+
+  // TODO: reduce the large numbers of whole-array iterations by combinging the array methods used
+  // here into a big for-loop
+  protected getValuesFromResults<ResultItem>(
+    inputData: ResultItem[],
+    options?: GetValuesFromResultsOptions
+  ): ResultItem[] {
+    const { primaryKeysMap, schema } = this;
+    const { isArray, nestedObjectContainerPath } = schema;
+    const { filters, flattenArray, hasNonPrimaryKeyFilters, primaryKeyFiltersToForceCheck } = options || {};
+    let initialResults = [...inputData];
+    if (nestedObjectContainerPath) {
+      initialResults = initialResults.map(item => {
+        if (item && typeof item === 'object' && !(item instanceof Date)) {
+          return getNested(item, nestedObjectContainerPath, { removeNestedFieldEscapeSign: true }).unifiedValue;
+        }
+        return item;
+      }) as ResultItem[];
+    }
+    if (isArray && (flattenArray || typeof flattenArray === 'undefined')) {
+      initialResults = initialResults.flat() as ResultItem[];
+    }
+    if (!hasNonPrimaryKeyFilters || !filters) {
+      return initialResults.filter(item => typeof item !== 'undefined' && item !== null);
+    }
+    // filter by the results' object data
+    return initialResults.filter(item =>
+      this.filterItem<ResultItem>(item, filters, {
+        keysToSkip: primaryKeysMap,
+        skippableKeysToForceCheck: primaryKeyFiltersToForceCheck
+      })
+    );
+  }
+
   protected async prepare(data: Entity, options?: PrepareOptions): Promise<{ data: Entity; storeEntityKey: string }> {
     const { primaryKeys, schema, store, storeDelimiter } = this;
     const { columns, name: entityName, storeKey: entityStoreKey } = schema;
@@ -340,8 +382,10 @@ export class RedisRepositoryService<Entity> {
     const { generatePrimaryKeys, onConflict: optOnConflict, validate: optValidate } = opt;
     const onConflict = optOnConflict || SaveOptionsOnConflict.ThrowError;
     let allPKValuesExist = true;
-    let preparedData = ld.cloneDeep(data) as Record<string, unknown>;
+    let preparedData = ld.cloneDeep(data) as GenericObject<unknown>;
     let storeEntityKey = '';
+    // set up the construction of the store keys by primary keys
+    // additionally, perform the generation of primary keys, if doing a create opearation
     for (const columnName of primaryKeys) {
       const { generated, type } = columns[columnName];
       const value = preparedData[columnName];
@@ -397,26 +441,28 @@ export class RedisRepositoryService<Entity> {
       }
       storeEntityKey += `${value}${storeDelimiter}`;
     }
+    if (storeEntityKey.endsWith(storeDelimiter)) {
+      storeEntityKey = storeEntityKey.substring(0, storeEntityKey.length - storeDelimiter.length);
+    }
     if (optValidate) {
-      const validationErrors = await validate(entityName, data as Record<string, unknown>);
+      const validationErrors = await validate(entityName, data as GenericObject<unknown>);
       if (validationErrors.length) {
         throw new ApplicationError(
           `[RedisRepositoryService ${entityName}][Validation Error]: ${validationErrors.join('\n')}`
         );
       }
     }
-    // TODO: escape regex symbols
-    storeEntityKey = storeEntityKey.replace(new RegExp(`${storeDelimiter}$`), '');
+    // TODO: make this work using getValuesFromResults
     if (onConflict !== SaveOptionsOnConflict.DoNothing && allPKValuesExist) {
-      const existingValue = await store.get<string | undefined>(storeEntityKey, { parseToJSON: false });
-      if (existingValue) {
+      const hasValue = await store.get<string | undefined>(storeEntityKey, { withValues: false });
+      if (hasValue) {
         if (onConflict === SaveOptionsOnConflict.ThrowError) {
           throw new ApplicationError(
             `[RedisRepositoryService ${entityName}][Unique Error]: An entry already exists for key ${storeEntityKey}.`
           );
         }
         if (onConflict === SaveOptionsOnConflict.Update) {
-          const existingData = JSON.parse(existingValue) as Record<string, unknown>;
+          const existingData = await store.get<GenericObject<unknown>>(storeEntityKey, { parseToJSON: true });
           preparedData = ld.merge(existingData, preparedData);
         } else {
           throw new ApplicationError(
@@ -433,12 +479,21 @@ export class RedisRepositoryService<Entity> {
     data: Entity | Entity[],
     options?: SaveOptions
   ): Promise<ResultItem[]> {
-    const { defaultTTL, schema, store, storeDelimiter } = this;
-    const { storeKey: entityStoreKey } = schema;
-    const { delete: optDelete, onConflict, transactionId, ttl, validate } = options || ({} as SaveOptions);
+    const { defaultTTL, experimentalDeletionEnabled, innerPrimaryKeys, primaryKeysMap, schema, store, storeDelimiter } =
+      this;
+    const { isArray, nestedObjectContainerPath, storeKey: entityStoreKey } = schema;
+    const {
+      delete: optDelete,
+      generatePrimaryKeys,
+      onConflict,
+      transactionId,
+      ttl,
+      validate
+    } = options || ({} as SaveOptions);
     const actualData = data instanceof Array ? data : [data];
     if (optDelete) {
       const prepareOptions: PrepareOptions = {
+        generatePrimaryKeys: false,
         onConflict: SaveOptionsOnConflict.DoNothing,
         validate: false
       };
@@ -448,13 +503,76 @@ export class RedisRepositoryService<Entity> {
           `${entityStoreKey}${storeDelimiter}${(await this.prepare(actualData[i], prepareOptions)).storeEntityKey}`
         );
       }
+      // delete from arrays and nestedObjects;
+      // this is kind of a repeat of find, but with the key paths included and with dedicated
+      // filtering by inner primary keys
+      if (
+        experimentalDeletionEnabled &&
+        (isArray || nestedObjectContainerPath) &&
+        innerPrimaryKeys.length &&
+        deleteKeys.length
+      ) {
+        const results = await Promise.all(deleteKeys.map(key => store.get<ResultItem>(key, { parseToJSON: true })));
+        // const deletePromises: Promise<unknown>[] = [];
+        const newResults: ResultItem[] = [];
+        results.forEach((resultItem, resultItemIndex) => {
+          if (!resultItem || typeof resultItem !== 'object' || resultItem instanceof Date) {
+            newResults.push(resultItem);
+            return;
+          }
+          let innerPaths: string[] = [];
+          let innerValues: ResultItem[] = [];
+          if (nestedObjectContainerPath) {
+            const { paths, values } = getNested(resultItem, nestedObjectContainerPath, {
+              removeNestedFieldEscapeSign: true
+            });
+            innerPaths = paths;
+            innerValues = values as ResultItem[];
+            // TODO: combine with isArray
+          } else if (isArray) {
+            innerValues = resultItem as ResultItem[];
+          } else {
+            innerValues = [resultItem];
+          }
+          // TODO: complete this logic
+          innerValues.forEach((innerValue, innerValueIndex) => {
+            const shouldDelete = this.filterItem<ResultItem>(
+              innerValue,
+              actualData[resultItemIndex] as GenericObject<unknown>,
+              {
+                keysToSkip: primaryKeysMap
+              }
+            );
+            if (!shouldDelete) {
+              return;
+            }
+            if (innerPaths[innerValueIndex]) {
+              setNested(results[resultItemIndex], innerPaths[innerValueIndex], undefined, {
+                removeNestedFieldEscapeSign: true
+              });
+              return;
+            }
+            if (isArray) {
+              setNested(results, `${resultItemIndex}`, (resultItem as ResultItem[]).splice(innerValueIndex, 1), {
+                removeNestedFieldEscapeSign: true
+              });
+              return;
+            }
+            results.splice(resultItemIndex, 1);
+          });
+        });
+        return [];
+      }
+      // default use case - regular people storing regular objects in redis
       if (deleteKeys.length) {
         await store.delete(deleteKeys, { transactionId });
       }
       return deleteKeys as ResultItem[];
     }
+    // TODO: create and update in arrays and nestedObjects
+    // TODO: differenatiate between create and update based on generatePrimaryKeys
     const prepareOptions: PrepareOptions = {
-      generatePrimaryKeys: true,
+      generatePrimaryKeys,
       onConflict,
       validate
     };